ui: fixed issue related to profiler and data quality
