+++ openmetadata-ui/src/main/resources/ui/src/components/DataQuality/SummaryPannel/SummaryPanel.interface.ts
@@ -13,5 +13,6 @@ import { TestSummary } from 'generated/tests/testSuite';
  *  limitations under the License.
  */
 export interface SummaryPanelProps {
-  testSummary?: TestSummary;
+  testSummary: TestSummary;
+  isLoading?: boolean;
 }
