+++ openmetadata-ui/src/main/resources/ui/src/components/TableProfiler/Component/ProfilerSettingsModal.tsx
@@ -170,10 +170,10 @@ const ProfilerSettingsModal: React.FC<ProfilerSettingsModalProps> = ({
     form.setFieldsValue({
       profileSampleType,
       profileSamplePercentage: profileSampleTypeCheck
-        ? profileSample || 100
+        ? profileSample ?? 100
         : 100,
       profileSampleRows: !profileSampleTypeCheck
-        ? profileSample || 100
+        ? profileSample ?? 100
         : undefined,
     });
