+++ openmetadata-ui/src/main/resources/ui/src/pages/DataQuality/DataQualityPage.tsx
@@ -12,23 +12,36 @@
  */
 
 import { Col, Row, Tabs, TabsProps, Typography } from 'antd';
+import { AxiosError } from 'axios';
 import PageLayoutV1 from 'components/containers/PageLayoutV1';
 import { SummaryPanel } from 'components/DataQuality/SummaryPannel/SummaryPanel.component';
 import { TestCases } from 'components/DataQuality/TestCases/TestCases.component';
 import { TestSuites } from 'components/DataQuality/TestSuites/TestSuites.component';
+import { usePermissionProvider } from 'components/PermissionProvider/PermissionProvider';
 import TabsLabel from 'components/TabsLabel/TabsLabel.component';
-import React, { useMemo } from 'react';
+import { INITIAL_TEST_SUMMARY } from 'constants/TestSuite.constant';
+import { TestSummary } from 'generated/tests/testCase';
+import React, { useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useHistory, useParams } from 'react-router-dom';
+import { getTestCaseExecutionSummary } from 'rest/testAPI';
 import { getDataQualityPagePath } from 'utils/RouterUtils';
+import { showErrorToast } from 'utils/ToastUtils';
 import { DataQualityPageTabs } from './DataQualityPage.interface';
 
 const DataQualityPage = () => {
   const { t } = useTranslation();
   const { tab: activeTab } = useParams<{ tab: DataQualityPageTabs }>();
   const history = useHistory();
+  const [summary, setSummary] = useState<TestSummary>(INITIAL_TEST_SUMMARY);
+  const [isSummaryLoading, setIsSummaryLoading] = useState(true);
+  const { permissions } = usePermissionProvider();
+  const { testCase: testCasePermission } = permissions;
 
-  const summaryPanel = useMemo(() => <SummaryPanel />, []);
+  const summaryPanel = useMemo(
+    () => <SummaryPanel isLoading={isSummaryLoading} testSummary={summary} />,
+    [summary, isSummaryLoading]
+  );
 
   const tabDetails = useMemo(() => {
     const tab: TabsProps['items'] = [
@@ -74,6 +87,25 @@ const DataQualityPage = () => {
       history.push(getDataQualityPagePath(activeKey as DataQualityPageTabs));
     }
   };
+
+  const fetchTestSummary = async () => {
+    try {
+      const response = await getTestCaseExecutionSummary();
+      setSummary(response);
+    } catch (error) {
+      showErrorToast(error as AxiosError);
+    } finally {
+      setIsSummaryLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    if (testCasePermission?.ViewAll || testCasePermission?.ViewBasic) {
+      fetchTestSummary();
+    } else {
+      setIsSummaryLoading(false);
+    }
+  }, []);
 
   return (
     <PageLayoutV1 pageTitle="Quality">
